backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
chromadb==0.5.3
sentence-transformers==3.0.1
requests==2.32.3
yfinance==0.2.41
python-decouple==3.8
django-extensions==3.2.3
channels==4.1.0
daphne==4.1.2
